Updated to work with Windows 10 IoT Core 10.0.10531.  Added networking metrics to home page.

+++ WindowsIoTDashboard.App/Models/ProcessesModel.cs
@@ -6,14 +6,14 @@ namespace WindowsIoTDashboard.App.Models
     {
         public class Process
         {
-            public float CPUUsage { get; set; }
+            public double CPUUsage { get; set; }
             public string ImageName { get; set; }
-            public float PrivateWorkingSet { get; set; }
+            public double PrivateWorkingSet { get; set; }
             public int ProcessId { get; set; }
             public int SessionId { get; set; }
             public string UserName { get; set; }
-            public float VirtualSize { get; set; }
-            public float WorkingSetSize { get; set; }
+            public double VirtualSize { get; set; }
+            public double WorkingSetSize { get; set; }
             public string PackageFullName { get; set; }
             public override string ToString()
             {
